Fixed querying deleted docs
Now that deleted docs are in a separate table, they can't be queried
the way they were before. Instead, QueryParser now checks whether the
WHERE clause matches _only_ deleted docs; if so, it operates on the
kv_deleted table instead, and the `_deleted` property becomes a no-op
(evaluates to `true`.) Otherwise, any use of the `_deleted` property
is an error.
This is a bit more limited since you can't write a query that will
match both deleted and non-deleted docs, but it should handle the
typical use cases.